pytestUI/requirements.txt
allure-pytest==2.8.15
allure-python-commons==2.8.15
appdirs==1.4.4
APScheduler==3.6.3
atomicwrites==1.3.0
attrs==19.3.0
BeautifulReport==0.1.2
beautifulsoup4==4.8.1
black==19.10b0
certifi==2019.9.11
cffi==1.14.0
chardet==3.0.4
Click==7.0
colorama==0.4.3
colorlog==4.1.0
crypto==1.4.1
cryptography==2.9.2
filetype==1.0.7
Flask==1.1.1
har2case==0.3.1
HttpRunner==1.5.8
idna==2.8
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.10.3
jmespath==0.9.5
loguru==0.4.1
lxml==4.5.0
MarkupSafe==1.1.1
more-itertools==8.2.0
Naked==0.1.31
numpy==1.17.4
packaging==20.3
parameterized==0.7.1
ParamUnittest==0.2
pathspec==0.8.0
Pillow==6.2.1
pluggy==0.13.1
py==1.8.1
pycparser==2.20
pycryptodome==3.9.4
pydantic==1.5.1
Pygments==2.6.1
PyMySQL==0.9.3
pyparsing==2.4.7
Pypubsub==4.0.3
pytest==5.4.2
pytz==2019.3
PyYAML==5.1.2
regex==2020.5.14
requests==2.22.0
requests-toolbelt==0.9.1
robotframework==3.1.2
robotframework-ride==1.7.4.1
schedule==0.6.0
selenium==3.141.0
send-email==201212
shellescape==3.4.1
six==1.13.0
soupsieve==1.9.5
toml==0.10.1
typed-ast==1.4.1
tzlocal==2.0.0
urllib3==1.25.7
wcwidth==0.1.9
Werkzeug==0.16.0
win32-setctime==1.0.1
wxPython==4.0.7.post2
xlrd==1.2.0
zipp==3.1.0
